# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: DotDeck API
  version: 1.0.0
servers:
  - url: /api/v1
...
paths:
...
  /tags:
    get:
      summary: List tags
      tags: [Tags]
      parameters:
        - in: query
          name: all
          schema: { type: integer, enum: [0, 1] }
          description: Include unofficial tags when `1`.
      responses: { "200": { description: Tag array } }
...
  /admin/users/{id}/ban:
    post:
      summary: Ban a user permanently
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        description: Reason for ban
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "204":
          description: Banned
    delete:
      summary: Un‑ban a user
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Unbanned
  /admin/tags:
    post:
      summary: Create a new tag
      tags: [Admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, tagType]
              properties:
                name:
                  type: string
                tagType:
                  type: integer
                  description: FK to dotdeck_tag_type.id
      responses:
        "201":
          description: New tag id
  /admin/audit:
    get:
      summary: Recent audit rows latest 100
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Array of audit objects
  /admin/tags/{id}/approve:
    patch:
      summary: Approve a tag
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses: { "204": { description: Approved } }
  /admin/tags/{id}/merge-into/{targetId}:
    patch:
      summary: Merge two tags
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: path
          name: targetId
          required: true
          schema: { type: integer }
      responses: { "204": { description: Merged } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...